+++ tools/agents/hooks/stop-check.spec.ts
@@ -1,4 +1,5 @@
 // https://code.claude.com/docs/en/hooks#stop
+import { Json } from '../../../lib/util/schema-utils/index.ts';
 import { BlockOutput } from './utils/schemas.ts';
 
 const { exec } = vi.hoisted(() => ({ exec: vi.fn() }));
@@ -47,7 +48,7 @@ it('outputs block JSON when pnpm check --all fails', async () => {
   await import('./stop-check.ts');
 
   expect(consoleSpy).toHaveBeenCalledOnce();
-  const output = BlockOutput.parse(JSON.parse(consoleSpy.mock.calls[0][0]));
+  const output = Json.pipe(BlockOutput).parse(consoleSpy.mock.calls[0][0]);
   expect(output).toEqual({
     decision: 'block',
     reason:
